requirements.txt
Django==1.4.10
requests==1.2.3
redis==2.8.0
simplejson==3.3.1
djangorestframework==2.3.9
markdown==2.0.3
django-filter==0.7
isodate==0.4.9
python-dateutil==2.2
django-private-views==1.1
celery==3.1.7
django-celery==3.1.1
pygeocoder==1.2.2
colorama==0.2.7
django-taggit==0.11.2
django-extensions==1.3.3
django-push-notifications
django-uuidfield